convert JsdomWithTextEncoderEnvironment to typescript
